# image minifyBy2 faster, mask minifyBy2, path fills take blendMode