--- a/src/pixie/paths.nim
+++ b/src/pixie/paths.nim
@@ -360,6 +360,9 @@ proc rect*(path: var Path, x, y, w, h: float32) =
   path.lineTo(x, y + h)
   path.closePath()
 
+proc rect*(path: var Path, pos: Vec2, wh: Vec2) {.inline.} =
+  path.rect(pos.x, pos.y, wh.x, wh.y)
+
 proc polygon*(path: var Path, x, y, size: float32, sides: int) =
   ## Draws a n sided regular polygon at (x, y) with size.
   path.moveTo(x + size * cos(0.0), y + size * sin(0.0))
@@ -886,6 +889,7 @@ proc fillShapes(
   image: Image,
   shapes: seq[seq[Vec2]],
   color: ColorRGBA,
+  blendMode: BlendMode,
   windingRule: WindingRule
 ) =
   let (topHalf, bottomHalf, fullHeight) =
@@ -898,6 +902,10 @@ proc fillShapes(
     startX = max(0, bounds.x.int)
     startY = max(0, bounds.y.int)
     stopY = min(image.height, (bounds.y + bounds.h).int)
+    blender = blendMode.blenderPremultiplied()
+
+  when defined(amd64) and not defined(pixieNoSimd):
+    let blenderSimd = blendMode.blenderSimd()
 
   var
     coverages = newSeq[uint8](image.width)
@@ -979,10 +987,7 @@ proc fillShapes(
           let
             index = image.dataIndex(x, y)
             backdrop = mm_loadu_si128(image.data[index].addr)
-          mm_storeu_si128(
-            image.data[index].addr,
-            blendNormalPremultiplied(backdrop, source)
-          )
+          mm_storeu_si128(image.data[index].addr, blenderSimd(backdrop, source))
 
         x += 4
 
@@ -1003,7 +1008,7 @@ proc fillShapes(
           source.a = ((color.a.uint16 * coverage) div 255).uint8
 
         let backdrop = image.getRgbaUnsafe(x, y)
-        image.setRgbaUnsafe(x, y, blendNormalPremultiplied(backdrop, source))
+        image.setRgbaUnsafe(x, y, blender(backdrop, source))
       inc x
 
 proc fillShapes(
@@ -1160,35 +1165,38 @@ proc fillPath*(
   image: Image,
   path: SomePath,
   color: ColorRGBA,
+  blendMode = bmNormal,
   windingRule = wrNonZero
 ) {.inline.} =
-  image.fillShapes(parseSomePath(path), color, windingRule)
+  image.fillShapes(parseSomePath(path), color, blendMode, windingRule)
 
 proc fillPath*(
   image: Image,
   path: SomePath,
   color: ColorRGBA,
   pos: Vec2,
+  blendMode = bmNormal,
   windingRule = wrNonZero
 ) =
   var shapes = parseSomePath(path)
   for shape in shapes.mitems:
     for segment in shape.mitems:
       segment += pos
-  image.fillShapes(shapes, color, windingRule)
+  image.fillShapes(shapes, color, blendMode, windingRule)
 
 proc fillPath*(
   image: Image,
   path: SomePath,
   color: ColorRGBA,
   mat: Mat3,
+  blendMode = bmNormal,
   windingRule = wrNonZero
 ) =
   var shapes = parseSomePath(path)
   for shape in shapes.mitems:
     for segment in shape.mitems:
       segment = mat * segment
-  image.fillShapes(shapes, color, windingRule)
+  image.fillShapes(shapes, color, blendMode, windingRule)
 
 proc fillPath*(
   mask: Mask,
@@ -1226,6 +1234,7 @@ proc strokePath*(
   path: SomePath,
   color: ColorRGBA,
   strokeWidth = 1.0,
+  blendMode = bmNormal,
   windingRule = wrNonZero
 ) =
   let strokeShapes = strokeShapes(
@@ -1233,7 +1242,7 @@ proc strokePath*(
     strokeWidth,
     windingRule
   )
-  image.fillShapes(strokeShapes, color, windingRule)
+  image.fillShapes(strokeShapes, color, blendMode, windingRule)
 
 proc strokePath*(
   image: Image,
@@ -1241,6 +1250,7 @@ proc strokePath*(
   color: ColorRGBA,
   strokeWidth = 1.0,
   pos: Vec2,
+  blendMode = bmNormal,
   windingRule = wrNonZero
 ) =
   var strokeShapes = strokeShapes(
@@ -1251,7 +1261,7 @@ proc strokePath*(
   for shape in strokeShapes.mitems:
     for segment in shape.mitems:
       segment += pos
-  image.fillShapes(strokeShapes, color, windingRule)
+  image.fillShapes(strokeShapes, color, blendMode, windingRule)
 
 proc strokePath*(
   image: Image,
@@ -1259,6 +1269,7 @@ proc strokePath*(
   color: ColorRGBA,
   strokeWidth = 1.0,
   mat: Mat3,
+  blendMode = bmNormal,
   windingRule = wrNonZero
 ) =
   var strokeShapes = strokeShapes(
@@ -1269,7 +1280,7 @@ proc strokePath*(
   for shape in strokeShapes.mitems:
     for segment in shape.mitems:
       segment = mat * segment
-  image.fillShapes(strokeShapes, color, windingRule)
+  image.fillShapes(strokeShapes, color, blendMode, windingRule)
 
 proc strokePath*(
   mask: Mask,