--- a/tests/test_images.nim
+++ b/tests/test_images.nim
@@ -85,3 +85,9 @@ block:
     a = readImage("tests/images/flipped1.png")
     b = a.minifyBy2()
   b.writeFile("tests/images/minifiedBy2.png")
+
+block:
+  let
+    a = readImage("tests/images/flipped1.png")
+    b = a.minifyBy2(2)
+  b.writeFile("tests/images/minifiedBy4.png")