… tests/images/masks/maskMinified.png
… tests/images/minifiedBy2.png
… tests/images/minifiedBy4.png