before:
`minifyBy2 ......................... 12.260 ms     12.663 ms    ±0.302   x392`
after:
`minifyBy2 .......................... 3.782 ms      3.967 ms    ±0.271  x1000`